backend/requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
Pillow
requests